pyproject.toml
[project]
name = "agentic-rag"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "chromadb>=1.1.0",
    "fastapi>=0.116.2",
    "groq>=0.31.1",
    "langchain>=0.3.27",
    "langchain-chroma>=0.2.6",
    "langchain-community>=0.3.29",
    "langchain-huggingface>=0.3.1",
    "langchain-openai>=0.3.33",
    "langgraph>=0.6.7",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.0.0",
    "sentence-transformers>=5.1.0",
    "streamlit>=1.40.0",
    "uvicorn>=0.35.0",
]
